style: Dex페이지 1차 스타일링
