+++ src/components/Dashboard.jsx
@@ -1,4 +1,11 @@
-import PokemonCard from "./PokemonCard";
+import { Link } from "react-router-dom";
+import styled from "styled-components";
+
+const StDashboardUl = styled.ul`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
 
 const Dashboard = ({ selectedPokemon, setSelectedPokemon }) => {
   const pokemonBallCount = 6 - selectedPokemon.length;
@@ -9,11 +16,15 @@ const Dashboard = ({ selectedPokemon, setSelectedPokemon }) => {
 
   return (
     <header>
-      <h1>나만의 포켓몬 도감</h1>
-      <ul>
+      <StDashboardUl>
+        <h1>나만의 포켓몬 도감</h1>
         {selectedPokemon.map((pokemon) => (
           <li key={pokemon.id}>
-            <PokemonCard pokemon={pokemon} />
+            <Link to={`/detail?pokemon=${pokemon.id}`}>
+              <img src={pokemon.img_url} alt={pokemon.korean_name} />
+              <h3>{pokemon.korean_name}</h3>
+              <p>No.{pokemon.id.toString().padStart(3, "0")}</p>
+            </Link>
             <button onClick={() => cancelPokemon(pokemon.id)}>취소</button>
           </li>
         ))}
@@ -26,7 +37,7 @@ const Dashboard = ({ selectedPokemon, setSelectedPokemon }) => {
                 <img src="/image/monsterBall-img.png" alt="빈 포켓볼" />
               </li>
             ))}
-      </ul>
+      </StDashboardUl>
     </header>
   );
 };
