+++ src/components/GlobalStyle.jsx
@@ -39,6 +39,7 @@ const GlobalStyle = createGlobalStyle`
 
     #root {
         display: flex;
+        flex-direction: column;
         justify-content: center;
         align-items: center;
     }
